Pull ID type enum to the package level, and hide
