+++ base/src/test/java/io/spine/base/IdentifierTest.java
@@ -51,10 +51,10 @@ import org.junit.jupiter.api.Test;
 import static com.google.common.truth.Truth.assertThat;
 import static io.spine.base.Identifier.EMPTY_ID;
 import static io.spine.base.Identifier.NULL_ID;
-import static io.spine.base.Identifier.Type.INTEGER;
-import static io.spine.base.Identifier.Type.LONG;
-import static io.spine.base.Identifier.Type.MESSAGE;
-import static io.spine.base.Identifier.Type.STRING;
+import static io.spine.base.IdType.INTEGER;
+import static io.spine.base.IdType.LONG;
+import static io.spine.base.IdType.MESSAGE;
+import static io.spine.base.IdType.STRING;
 import static io.spine.base.Identifier.checkSupported;
 import static io.spine.base.Identifier.findField;
 import static io.spine.base.Identifier.newUuid;
